Move iterators back to system
* Move SortedMapIterator to iteratorsImpl
