+++ core/src/test/java/org/apache/accumulo/core/iterators/SortedMapIteratorTest.java
@@ -23,6 +23,7 @@ import java.util.TreeMap;
 import org.apache.accumulo.core.client.SampleNotPresentException;
 import org.apache.accumulo.core.client.sample.RowSampler;
 import org.apache.accumulo.core.client.sample.SamplerConfiguration;
+import org.apache.accumulo.core.iteratorsImpl.system.SortedMapIterator;
 import org.junit.Test;
 
 public class SortedMapIteratorTest {
