+++ server/tserver/src/main/java/org/apache/accumulo/tserver/MemKeyConversionIterator.java
@@ -26,10 +26,10 @@ import org.apache.accumulo.core.data.ByteSequence;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Range;
 import org.apache.accumulo.core.data.Value;
-import org.apache.accumulo.core.iterators.InterruptibleIterator;
 import org.apache.accumulo.core.iterators.IteratorEnvironment;
 import org.apache.accumulo.core.iterators.SortedKeyValueIterator;
 import org.apache.accumulo.core.iterators.WrappingIterator;
+import org.apache.accumulo.core.iteratorsImpl.system.InterruptibleIterator;
 
 class MemKeyConversionIterator extends WrappingIterator implements InterruptibleIterator {
   private MemKey currKey = null;
